fix package organization
