+++ cmd/dxworker/config.go
@@ -7,15 +7,13 @@ import (
 	"github.com/kudarap/dotagiftx/steam"
 )
 
-type (
-	Config struct {
-		SigKey      string
-		Prod        bool
-		Addr        string
-		SpanEnabled bool `envconfig:"SPAN_ENABLED"`
-		Rethink     rethink.Config
-		Redis       redis.Config
-		Steam       steam.Config
-		Log         log.Config
-	}
-)
+type Config struct {
+	SigKey      string
+	Prod        bool
+	Addr        string
+	SpanEnabled bool `envconfig:"SPAN_ENABLED"`
+	Rethink     rethink.Config
+	Redis       redis.Config
+	Steam       steam.Config
+	Log         log.Config
+}
