+++ tools/inventoryparser/main.go
