+++ tools/marketmocker/main.go
