+++ tools/runonce/bid.go
@@ -1,4 +1,4 @@
-package fixes
+package runonce
 
 import (
 	"context"
