+++ tools/runonce/fixes.go
@@ -1,9 +1,7 @@
-package fixes
+package runonce
 
 import (
-	"context"
 	"log"
-	"math/rand"
 	"strings"
 
 	dgx "github.com/kudarap/dotagiftx"
@@ -19,27 +17,6 @@ func ReIndexAll(
 			log.Println("err", err)
 		}
 
-	}
-}
-
-func GenerateFakeMarket(
-	itemStg dgx.ItemStorage,
-	userStg dgx.UserStorage,
-	marketSvc dgx.MarketService,
-) {
-
-	ctx := context.Background()
-	ii, _ := itemStg.Find(dgx.FindOpts{})
-	uu, _ := userStg.Find(dgx.FindOpts{})
-	for _, item := range ii {
-		for _, user := range uu {
-			m := &dgx.Market{
-				ItemID: item.ID,
-				Price:  float64(rand.Intn(1000)) / 10,
-			}
-			auc := dgx.AuthToContext(ctx, &dgx.Auth{UserID: user.ID})
-			marketSvc.Create(auc, m)
-		}
 	}
 }
